file sorter

+++ core/src/main/java/com/rainyalley/architecture/core/arithmetic/sort/FileSorter.java
@@ -1,13 +1,19 @@
-package com.rainyalley.architecture.util;
+package com.rainyalley.architecture.core.arithmetic.sort;
 
+import com.rainyalley.architecture.util.Chunk;
+import com.rainyalley.architecture.util.QueuedLineReader;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.io.FileUtils;
 import org.springframework.util.Assert;
 
 import java.io.*;
 import java.util.*;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+/**
+ * @author bin.zhang
+ */
 public class FileSorter {
 
 
@@ -18,15 +24,19 @@ public class FileSorter {
     private static final int INITIAL_CHUNK_LEVEL = 1;
 
     /**
+     * 第一行的行号
+     */
+    private static final int INITIAL_ROW_NUM = 1;
+
+    /**
      * io缓冲大小
      */
     private int ioBufferSize = 1024 * 1024 * 2;
 
-
     /**
      * 临时目录
      */
-    private String tmpDir = "/var/tmp/fileCompare/";
+    private File tmpDir;
 
     /**
      * 行内容比较器
@@ -41,29 +51,70 @@ public class FileSorter {
     /**
      * 每个块的行数
      */
-    private int initialChunkSize = 10000;
+    private int initialChunkSize = 1;
 
+    /**
+     *
+     */
+    private boolean clean;
 
-    public FileSorter(Comparator<String> comparator) {
+    /**
+     *
+     * @param comparator 行比较器
+     * @param mergeWayNum 归并路数
+     * @param initialChunkSize 初始块中的行数
+     * @param ioBufferSize io缓冲区容量
+     * @param tmpDir 临时目录
+     * @param clean 排序完成后，是否清除临时文件
+     */
+    public FileSorter(Comparator<String> comparator, int mergeWayNum, int initialChunkSize, int ioBufferSize, File tmpDir, boolean clean) {
+        Assert.isTrue(mergeWayNum >=2, "mergeWayNum must greater than 2");
+        Assert.isTrue(initialChunkSize >=1, "initialChunkSize must greater than 1");
+        Assert.isTrue(ioBufferSize >=1024, "ioBufferSize must greater than 1024");
+        Assert.notNull(tmpDir, "tmpDir can not be null");
+
         this.comparator = comparator;
+        if(!tmpDir.exists()){
+            tmpDir.mkdirs();
+        }
+
+        this.mergeWayNum = mergeWayNum;
+        this.initialChunkSize = initialChunkSize;
+        this.ioBufferSize = ioBufferSize;
+        this.tmpDir = tmpDir;
+        this.clean = clean;
     }
 
-    public void sort(File file) throws IOException {
-        List<Chunk> splitChunkList = split(file);
+
+    /**
+     * 排序
+     * @param original 原文件
+     * @param dest 排序后的文件
+     * @throws IOException
+     */
+    public void sort(File original, File dest) throws IOException {
+        Assert.isTrue(original.exists(), String.format("original file not fond: %s", original.getPath()));
+        Assert.isTrue(!dest.exists(), String.format("dest file must not exist: %s", dest.getPath()));
+
+        File workDir = workDir(original);
+        if(!workDir.exists()){
+            workDir.mkdir();
+        } else {
+            FileUtils.cleanDirectory(workDir);
+        }
+
+        List<Chunk> splitChunkList = split(original);
 
         Queue<Chunk> chunkQueue = new LinkedList<>(splitChunkList);
         int currentLevel = INITIAL_CHUNK_LEVEL;
-
         while (true){
             List<Chunk> pollChunks =  pollChunks(chunkQueue, currentLevel);
             if(CollectionUtils.isEmpty(pollChunks)){
                 currentLevel++;
                 continue;
             }
-
             //合并
-            Chunk mergedChunk = merge(pollChunks, file);
-
+            Chunk mergedChunk = merge(pollChunks, original);
             //chunkQueue 中没有后续的元素，表示此次合并是最终合并
             if(chunkQueue.size() == 0){
                 chunkQueue.add(mergedChunk);
@@ -74,8 +125,15 @@ public class FileSorter {
         }
 
         Chunk finalChunk = chunkQueue.poll();
-        File finalChunkFile = chunkFile(finalChunk, file);
-        System.out.println(String.format("final chunk: %s", finalChunk.toString()));
+        File finalChunkFile = chunkFile(finalChunk, original);
+        boolean renamed =  finalChunkFile.renameTo(dest);
+        if(!renamed){
+            throw new IllegalStateException(String.format("rename failure: %s >>> %s", finalChunkFile.getPath(), dest.getPath()));
+        }
+
+        if(clean){
+            FileUtils.deleteDirectory(workDir);
+        }
     }
 
     private Chunk merge(List<Chunk> chunks, File originalFile) throws IOException{
@@ -89,16 +147,16 @@ public class FileSorter {
         File mergedChunkFile = chunkFile(mergedChunk, originalFile);
 
         try(BufferedWriter mergedChunkFileWriter = new BufferedWriter(new FileWriter(mergedChunkFile))){
-            Stream<QueuedLineReader> qlrs = chunks.stream().map(c -> {
+            List<QueuedLineReader> qlrList = chunks.stream().map(c -> {
                 try {
                     return new QueuedLineReader(new BufferedReader(new FileReader(chunkFile(c, originalFile))));
                 } catch (FileNotFoundException e) {
                     throw new IllegalStateException(e);
                 }
-            });
+            }).collect(Collectors.toList());
 
             while (true){
-                Optional<QueuedLineReader> minReader = qlrs
+                Optional<QueuedLineReader> minReader = qlrList.stream()
                         .filter(p -> p.peek() != null)
                         .min((p,n) -> comparator.compare(p.peek(), n.peek())
                 );
@@ -115,7 +173,6 @@ public class FileSorter {
 
         return mergedChunk;
     }
-
 
 
     /**
@@ -148,18 +205,16 @@ public class FileSorter {
      * @throws IOException
      */
     private List<Chunk> split(File file) throws IOException{
-        if(!file.exists()){
-            throw new FileNotFoundException(String.format("file not fond %s", file.getPath()));
-        }
 
         List<Chunk> chunkList = new ArrayList<>();
 
         try(BufferedReader br = new BufferedReader(new FileReader(file), ioBufferSize)){
-            int idx = 0;
+            int rowNum = INITIAL_ROW_NUM;
             String line = null;
             List<String> chunkRows = new ArrayList<>(initialChunkSize);
 
             while (true){
+
                 line = br.readLine();
 
                 if(line != null){
@@ -168,10 +223,10 @@ public class FileSorter {
 
                 if(line == null || chunkRows.size() >= initialChunkSize){
                     chunkRows.sort(comparator);
-                    Chunk chunk = initialChunk(idx, chunkRows, file);
+                    Chunk chunk = initialChunk(rowNum, chunkRows, file);
                     chunkList.add(chunk);
+                    rowNum += chunkRows.size();
                     chunkRows.clear();
-                    idx += chunkRows.size();
                 }
 
                 if(line == null){
@@ -184,16 +239,18 @@ public class FileSorter {
 
     /**
      * 创建一个块
-     * @param idx 块首行的索引号
+     * @param rowNum 块首行的行号
      * @param chunkRows 块中的行数
      * @param originalFile 原始文件
      * @return
      * @throws IOException
      */
-    private Chunk initialChunk(int idx, List<String> chunkRows, File originalFile) throws IOException{
-        Chunk chunk = new Chunk(INITIAL_CHUNK_LEVEL, idx, chunkRows.size());
+    private Chunk initialChunk(int rowNum, List<String> chunkRows, File originalFile) throws IOException{
+        Chunk chunk = new Chunk(INITIAL_CHUNK_LEVEL, rowNum, chunkRows.size());
         File chunkFile = chunkFile(chunk, originalFile);
-        FileUtils.deleteQuietly(chunkFile);
+        if(chunkFile.exists()){
+            FileUtils.deleteQuietly(chunkFile);
+        }
         Assert.isTrue(chunkFile.createNewFile(), String.format("createNewFile failure: %s", chunkFile.getPath()));
         try(BufferedWriter bw = new BufferedWriter(new FileWriter(chunkFile), ioBufferSize)){
             for (String chunkRow : chunkRows) {
@@ -205,7 +262,11 @@ public class FileSorter {
     }
 
     private File chunkFile(Chunk chunk, File originalFile){
-        return new File(String.format(tmpDir + "/%s/%s_%s_%s.txt", originalFile.getName(), chunk.getLevel(), chunk.getIdx(), chunk.getSize()).replace("/", File.separator));
+        return new File(String.format(workDir(originalFile).getPath() + File.separator +"%s_%s_%s.txt", chunk.getLevel(), chunk.getIdx(), chunk.getSize()));
+    }
+
+    private File workDir(File originalFile){
+        return new File(tmpDir.getPath() + File.separator + originalFile.getName());
     }
 
 }
